feat: change bar style and move onclose logic to component
