+++ styles/globals.scss
@@ -120,6 +120,7 @@ body {
 }
 #notification {
     animation: slide-in 1s ease-in-out normal;
+    overflow: hidden;
 }
 
 @keyframes slide-in {
@@ -137,9 +138,11 @@ body {
     content: '';
     display: block;
     width: 100%;
-    height: 0.4rem;
+    height: 0.2rem;
     position: absolute;
     bottom: 0;
+    margin-bottom:-0.05rem;
+    opacity: 0.2;
     left: 0;
     background-color: rgba(0, 0, 0, 0.253);
     z-index: 1;
